backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-socketio==5.10.0
python-jose[cryptography]==3.3.0
requests==2.31.0
